[Cassandra LegoCastle 4/N] Run test in parallel
Summary:
Run test in parallel so that we could shorten the time spend to run the tests.
In specific, we provide ant with `test.runners` option along with the number of cores.

I'm not satisfied that we can't directly download the test report, but it seems there are not easy way to achieve that, we can improve this experience later on.

Test Plan:
Sandcastle.
The test could be finished within 12 minutes (used to be ~1-2 hours)!

Reviewers: fdeliege, #ig-cassandra, dikang

Reviewed By: dikang

Subscribers: dikang, trunkagent

Differential Revision: https://phabricator.intern.facebook.com/D3288731

Tasks: 10793373

Signature: t1:3288731:1463613305:1bab83ab1d812e403d3a75f6f667c86e2fdecf44