[stream] fix coruppted cell bug
Summary:
There is a subtle bug only reproduce in data mismatch scenario, which turns out the Mutation is used not only in rocksdb but in elsewhile, which probably is the cache:
when we write, we write the Mutation to rocksdb, and also keep the cells of mutation in the cache.
without this patch the cell.value() got flushed into rocksdb and becomes empty, thus every following read become empty.
with this patch we fixed the issue.

Test Plan: use logging to monitor the difference before and after this patch, noticed the cell got right value.

Reviewers: #ig-cassandra, wpc

Reviewed By: wpc

Subscribers: wpc, trunkagent, #coreappwatch, fdeliege

Differential Revision: https://phabricator.intern.facebook.com/D5325381

Signature: t1:5325381:1498622549:421fbdb37cf50be962b2ddd668712ae6a9ec2385