[rocksdb] Better progress info for netstats
Summary:
`nodetool netstats` doesn't show correct information for rocksdb.
One issue is `ProgressInfo.isComplete`  relies on receivedbytes/totalbytes, which is addressed in this diff.
Another issue is that the percentage of file streamed is not correct because estimated size from rocksdb is not quite right,  instead we are showing number of streamed pairs / estimated keys.

Test Plan:
1. pass unittest
2. test in stagecluster

Reviewers: #ig-cassandra, dikang

Subscribers: dikang, fdeliege, trunkagent

Differential Revision: https://phabricator.intern.facebook.com/D5818884