Prevent the rocksDB.getProperty() calls after RocksDBCF.close() is invoked
Summary:
When we drain Cassandra, which means stop any traffic to the Cassandra server, we want to close RocksDB instance as well,
to cancel any background job like compaction in RocksDB.

One challenge is that, after we close RocksDB, there should be no call to rocksdb handler, in our case, we need to prevent
the rocksDB.getProperty() calls, which are triggered by JMX agent running on the host. Otherwise, JVM may crash.

The change here is to add a flag in RocksDBCF to indicate whether closed() is called or not. If it is called, getProperty()
will simply return empty ArrayList.

Test Plan:
Added unit test testGetPropertyAfterClose() in RocksDBCFTest.java. It verified that after RocksDBCF.close() is invoked,
RocksDBCF.getProperty(). Also I verified that without the change in RocksDBCF.java, this unit test will crash.

Reviewers: dikang

Subscribers: raychen

Differential Revision: https://phabricator.intern.facebook.com/D6573246