Plugin Rocksandra read and write paths
Summary: When writing to a table, the rocksdb index table is also updated. When reading from a table, the data is read from the rocksdb index table.

Test Plan: Run RocksDBEngineTest to test the proper read and write of index. Run BasicSelectTest and SelectTest to test original functionality.

Reviewers: wpc, dikang, sdev

Differential Revision: https://phabricator.intern.facebook.com/D6826101

Tasks: T25257824