Use rocksdb when deleting from index
Summary: When deleting a row, the index table was not being updated. It only seemed to be working because of the way cassandra secondary indexes work. Dead rows would be returned from the index table, and this data would be used to search the data table. Since the data table doesn't have these dead rows, the result is what we expect.

Test Plan: RocksandraIndexTest.deleteRowsIndexTest

Reviewers: wpc, dikang, sdev

Differential Revision: https://phabricator.intern.facebook.com/D7474314

Tasks: T27751514