[ig][c*] Fix snapshot deletion on rocksandra
Summary: C* tries to infer keyspace names based on folder names and since we include the RocksDB data directory within the C* data directory, it tries to clear snapshots for the "rocksdbsharded" keyspace. This makes us look into the schema for keyspaces we are aware about and should operate on. This also fixes an issue with trying to delete a non-existing directory when attemping to repeatedly clean all snapshots.