Validate the need to update topology for gossiped peer before doing so
Summary:
Processing gossip messages can lead to invalidations of the TokenMap cache. The cache will be repopulated when there are subsequent read/write requests, since they need to determine the correct replica(s) to send the request to. The theory is the combination of constant cache invalidations + read/write reqs is causing the high latency in TokenMap repopulation.

This change will prevent unnecessary invalidations of the the TokenMap cache by checking the need to update a peer's rack/dc info before doing so.