Add health check during gossip to self-quarantine if unreachable from gossip
Summary:
When a [proxy] node loses network connection, it becomes partitioned from the rest of the ring. Service router will retry requests to the host, and ultimately mark down the host. The issue occurs when the host regains network connection. Since our tupperware health check is unable to disable the host in SMC, traffic continues to be sent to the host while it is unreachable. If the host regains network connection (e.g. after a short network outage), traffic is immediately sent to the C* node. The node's state of the world is one in which it is unable to communicate with its neighbors, and upon immediately receiving traffic it throws UnavailableExceptions. Once the node has gossiped with its neighbors, its view of the world is restored and is once again able to successfully find replicas and serve requests.

This change will detect when a node thinks all its neighbors are unreachable, and shut off all client ports (thrift, cql) until it regains a healthy view of the world through gossip.