Add a cpuLoad CQL function
Summary:
Add a new built-in CQL function `cpuLoad()`. This function is meant to be used by clients for as an efficient and scalable way to retrieve the CPU load on each Cassandra node for load balancing purpose.

This function cannot be implemented with a UDF because of package restrictions.

Test Plan:
The returned CPU Load seems to be roughly equal to the load reported by command line tools and Activity Monitor.

```
cqlsh> SELECT cpuLoad() FROM system.local;

 system.cpuload()
------------------
               12
```