Add a cassandraConfig CQL function
Summary:
Add a new built-in CQL function `cassandraConfig()`. This function is meant to be used to remotely inspect the Cassandra configuration as it has been loaded from `cassandra.yaml`. Some expected use cases:
* Access the source of truth of the current state of the configuration of a Cassandra instance.
* Inspect which nodes have been restarted and which haven't after a configuration change.
* Create more efficient / simpler to write tools that can rapidly introspect the configuration of 1000s of nodes through CQL.

Note that only the configuration properties from `cassandra.yaml` are returned here. The environment and JVM options are handled in D16403799.