Filter out high severity node during bootstrap streaming
Summary:
During a new DC rebuilding, we don't want to stream data from the rebuilding
DC. As the new DC may not have all data. Because of that, we set severity high
to avoid the read traffic. This fix filter out high severity node for
bootstrap, which includes node replacement and new node expansion.